set up inst-jobs
(why inst-jobs as opposed to something like sidekiq or the original
delayed_job? because it has mechanisms to serialize jobs and to
debounce multiple firings of jobs that are otherwise idempotent,
which makes it a lot easier to write things that do things like
synchronize users with stripe subscriptions without the risk of
two concurrent synchronization runs colliding with each other.
no other queue system for ruby that I'm aware of has those out of
the box.)