add a skeleton api to fetch badge tokens accessible by a given reader
right now it returns all badge tokens in the system, effectively
allowing access to anyone with a badge. the logic to limit the
badge tokens to only users who should be allowed to access the
badge reader in question is yet to come.

also prevent multiple users from sharing the same badge token;
that's not something that should happen in practice since they
will be randomly generated, but prevent against it happening
just in case.