allow deferring association of a certification issuance to a user
some certifications are taken by users for whom we haven't
collected an email address or otherwise know who they are and who
aren't members yet. (it's debatable whether that's a thing we
should do or whether we should require email addresses from all
recipients, but that's a topic for another time - @rioangelo and i
are agreeing to disagree on that :) provide a way for certifications
to be entered for such users by allowing a tenative name to be
specified in lieu of a user. the certification carries no privileges
until the user is retroactively connected later.