store a badge's id, token, and authorized status on scan
the id is a string burned into every badge. it's not secure to use
as authentication material because someone could lift someone
else's badge id by walking near them and then build a device that
acts like a badge with that id, so we continue to use the badge
token to perform authentication - but it's useful information to
have if someone is failing to scan in repeatedly and we're not
sure why.