add a control to allow a badge reader to be manually opened
this adds a link on the badge reader admin page that can be clicked
to request that the badge reader be opened for a moment immediately.
useful for when someone locks themselves out and calls upon an
admin to help. (hopefully the admin looks at the security cameras
first to make sure it's a legit ask... the social engineering
aspects of this are concerning, but the alternative is worse.)

in the process, set up action cable. this will enable many cool
things going forward, like being able to proactively let badge
readers know when their access list has changed.

(note that redis is now required to run heimdall in local dev
environments - this is needed so that the local dev server, the
local dev inst_jobs worker pool, and any local rails consoles all
talk nicely to each other when a background job or something on
the rails console fires off an action cable message.)

also document Api::ApiController a bit more, and ignore VSCode IDE
artifacts.