port synchrotron into heimdall
https://github.com/makesaltlake/synchrotron is a bot that posts
membership and financial statistics to the MSL leadership when
membership counts and other things change. With this commit,
Heimdall now has the ability to post Synchrotron's reports with
exactly the same formatting and in response to exactly the same
events.

With that, it should be possible to wind down Synchrotron and have
one less service to manage.