round the time in the subscription model test
somehow, on CI only (not on my local development box), the time
gets rounded as it makes its way through from the subscription to
the user's `subscription_created`, the result of which is that line
19 in the spec fails because normal_user.subscription_created and
subscription.started_at are off by however many microseconds
subscription.started_at contains. round the time to fix this.