sync waivers immediately into heimdall the moment they're signed
waivers will now be pushed into heimdall in response to a webhook
push. note that the webhook will need to be added manually to each
template's settings; make salt lake only has one waiver template,
so that's easy enough to do.

and, now that that's the case, revert back to doing a full waiver
import weekly instead of daily (which hopefully won't find any
changes).