whoops, missed a few uses of `subscription.plan`. also i think i know what's up
it looks like the one user who's running into this has multiple plans on
their subscription for some reason. i think stripe returns an
empty plan when we ask for *the* plan in such a case because there
are multiple, so it wouldn't be accurate to pick just one.

how to handle this is an open question - combining the two might
be the right approach, or treating each pricing as its own
subscription, or adding a additional model for pricings. (we don't
want to go too far off the beaten path if the information isn't
actually useful for heimdall though.)